scripts: build_flashable_zip: introduce custom install script

+++ scripts/internal/build_flashable_zip.sh
@@ -580,6 +580,14 @@ if [ -d "$SRC_DIR/target/$TARGET_CODENAME/installer/root" ]; then
     EVAL "cp -a \"$SRC_DIR/target/$TARGET_CODENAME/installer/root/\"* \"$TMP_DIR\"" || exit 1
 fi
 
+if [ -f "$SRC_DIR/target/$TARGET_CODENAME/installer/customize.sh" ]; then
+    LOG_STEP_IN "- Running target custom install script"
+    (
+    . "$SRC_DIR/target/$TARGET_CODENAME/installer/customize.sh"
+    ) || exit 1
+    LOG_STEP_OUT
+fi
+
 LOG "- Creating zip"
 EVAL "rm -f \"$TMP_DIR/rom.zip\"" || exit 1
 # https://android.googlesource.com/platform/build/+/refs/tags/android-15.0.0_r1/tools/releasetools/common.py#3601
